src/python/requirements.txt
colorzero==2.0
gpiozero==2.0.1
greenlet==3.1.1
importlib-metadata==4.13.0
importlib-resources==5.13.0
pillow==11.0.0
playwright==1.49.1
pyee==12.0.0
RPi.GPIO==0.7.1
spidev==3.6
typing-extensions==4.12.2
waveshare-epaper==1.3.0
zipp==3.21.0
